Found issue with target variables
Realized that I have a problem with the bottom 20% of revenue numbers from the dataset that was pulled. It didn't initially show up as a result of how I binned the data in a histogram for data wrangling.

Have decided to use the IMDB API package IMDbPy to pull accurate revenue information.

Will add a notebook showing my work on accessing that revenue information then I plan to read it into the data wrangling notebook as a .csv