Realized was taking exponential of predicted values incorrectly
Was getting scores that are not correct because I was taking the exponential of  MAE/Residuals calculated based on log-transformed true values and predicted values.  Needed to take exponential of true & predicted values THEN calculated scores.

Created a V 2.0 of preprocessing notebook when I first started using a log transformation to test metrics and see if I can find something else to do to solve this